% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{access_time}
\alias{access_time}
\title{Access Time}
\format{An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 1.}
\usage{
access_time
}
\description{
Access Time of datasets.
}
\keyword{datasets}
