% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmgm.r
\docType{package}
\name{rmgm-package}
\alias{rmgm-package}
\title{Climate Datasets For Turkey}
\source{
Source: \url{http://www.mgm.gov.tr}
}
\description{
\code{rmgm} provides climate datasets for \code{81} cities in Turkey. You
also can translate data from/to English or Turkish by \link{translate}
method.
}
\details{
\describe{
  \item{\link{trcli}}{Monthly climatological means and extremes}
  \item{\link{trexm}}{Recorded maximum/minimum temperatures and dates for each month}
  \item{\link{trex}}{Recorded extremes and dates}
}
}
\author{
İsmail SEZEN <sezenismail@gmail.com>
}
