% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.r
\name{translate}
\alias{translate}
\alias{translate.data.frame}
\title{Translate datasets}
\usage{
translate(x, lang = "tr")

\method{translate}{data.frame}(x, lang = "tr")
}
\arguments{
\item{x}{A \link{data.frame} to translate}

\item{lang}{Language to translate (\code{tr} or \code{en}).}
}
\description{
Translate \link{trcli}, \link{trex} or \link{trexm} data.frames to/from
English or Turkish.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Default method for \code{data.frame}
}}

