% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{trcli}
\alias{trcli}
\title{Monthly Climatological Means}
\format{An object of class \code{data.frame} with 648 rows and 16 columns.}
\source{
Source: \url{http://www.mgm.gov.tr}
}
\usage{
trcli
}
\description{
This dataset contains monthly climatological means and extremes for
\code{81} cities in Turkey.
}
\details{
Data contains following parameters:
\enumerate{
  \item{Monthly Total Precipitation Mean}{ (\code{mm})}
  \item{Minimum Temperature}{ (\code{degree Celcius})}
  \item{Maximum Temperature}{ (\code{degree Celcius})}
  \item{Mean of Minimum Temperatures}{ (\code{degree Celcius})}
  \item{Mean of Maximum Temperatures}{ (\code{degree Celcius})}
  \item{Average Sunshine Duration}{ (\code{hour})}
  \item{Average Temperature}{ (\code{degree Celcius})}
  \item{Mean of precipitation days}{ (\code{day})}
}

Column Names:
\enumerate{
  \item{City}{ (\code{Factor})}
  \item{Parameter}{ (\code{factor})}
  \item{Unit}{ (\code{factor})}
  \item{January}{ (\code{numeric})}
  \item{February}{ (\code{numeric})}
  \item{March}{ (\code{numeric})}
  \item{April}{ (\code{numeric})}
  \item{May}{ (\code{factor})}
  \item{June}{ (\code{numeric})}
  \item{July}{ (\code{numeric})}
  \item{August}{ (\code{numeric})}
  \item{September}{ (\code{numeric})}
  \item{October}{ (\code{numeric})}
  \item{November}{ (\code{numeric})}
  \item{December}{ (\code{numeric})}
  \item{Annual}{ (\code{numeric})}
}

Also you can obtain oberservation period by \code{period} attribute attached
to the data.
}
\seealso{
\link{translate}
}
\keyword{datasets}
