% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{trex}
\alias{trex}
\title{Extremes and dates}
\format{An object of class \code{data.frame} with 405 rows and 5 columns.}
\source{
Source: \url{http://www.mgm.gov.tr}
}
\usage{
trex
}
\description{
Data contains minimum & maximum recorded temperatures, maximum total daily
precipitation amount, maximum daily wind speed and maxmimum depth of snow
for 81 cities in Turkey.
}
\details{
Data contains following parameters:
\enumerate{
  \item{Minimum Temperature}{ (\code{degree Celcius})}
  \item{Maximum Temperature}{ (\code{degree Celcius})}
  \item{Maximum Snow Depth}{ (\code{cm})}
  \item{Maximum Daily Wind Speed}{ (\code{km/h})}
  \item{Maximum of Daily Total Precipitation Amount}{ (\code{mm})}
}

Column Names:
\enumerate{
  \item{City}{ (\code{Factor})}
  \item{Parameter}{ (\code{factor})}
  \item{Unit}{ (\code{factor})}
  \item{Date}{ (\code{date})}
  \item{Value}{ (\code{numeric})}
}
}
\seealso{
\link{translate}
}
\keyword{datasets}
