% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{trexm}
\alias{trexm}
\title{Monthly Extreme temperatures and dates}
\format{An object of class \code{data.frame} with 1944 rows and 4 columns.}
\source{
Source: \url{http://www.mgm.gov.tr}
}
\usage{
trexm
}
\description{
Data contains monthly recorded minimum and maximum temperatures for \code{81}
cities in Turkey.
}
\details{
Column Names:
\enumerate{
  \item{City}{ (\code{Factor})}
  \item{Parameter}{ (\code{factor})}
  \item{Date}{ (\code{date})}
  \item{Value}{ (\code{numeric})}
}

Also you can obtain oberservation period by \code{period} attribute attached
to the data.
}
\seealso{
\link{translate}
}
\keyword{datasets}
